Draft Homepage and Admin page
Including short description and Input fields for the name. On Sumbit
the input will be checked against a list of previous names and respond
with an error if the name already exists, otherwise add the name to the
name list and save it in a .json file on the server.
The Admin page contains a reset button, which will reset the names.json
to an empty array.